Move project structure to a more traditional setup.
This makes it so that we only have one gradle project in the repository.

There are a couple of advantages to this:
* This makes it easier to open the project in intellij making it easier to
contribute. You don't have to open up multiple projects anymore, the
standard VCS -> Checkout from version control -> git should work.
Or just opening the root project will work.
* Less files in the repository.
* Common configuration such as spotless is more easily and automatically
applied to subprojects.
* Easier to share code and introduce dependencies between projects.
* Easier to assemble any module since anything can now be build from
root.
* Tests and checks can be easily run for the entire repository with one
command.

This reduces the number of symlinks