feat: Add support for "fail fast" feature on FTL (#1370)
Fixes (N/A)

## Test Plan
> How do we know the code works?

Ran a locally built jar in CI, verified that this flag is behaving as expected.

## Checklist

- [X] Documented
- [X] Unit tested

## Background
When a matrix hits an infrastructure errors, it is automatically retried. This can cause issues in time sensitive scenarios, as results can take ~5-15m longer to report after an infrastructure error occurs. This PR exposes this flag via the CLI.