fix: Fix flank freezes when large number of matrices is launched (#2026)
Fixes #2021 

## Test Plan
> How do we know the code works?

It is impossible to test whole flow in sensible and reliable way. It would require mocking lots of methods and since the problem is related with multithreading I think it's not the best idea. What's more, test might be unreliable due to differences in CPU configurations on different machines.

@asadsalman was asked to verify fix, as a proof (thanks!)